Extract date/time builtins from builtin.c into builtin_date.c
builtin.c was 9,086 lines. Extract the self-contained date/time
subsystem into its own compilation unit:

- builtin_date.c (1,365 lines): time, microtime, getdate,
  gettimeofday, date, strftime, gmdate, localtime, idate,
  mktime/gmmktime, plus DateFormat and PH7_Strftime helpers

builtin.c is now 7,712 lines. Guarded with PH7_DISABLE_BUILTIN_FUNC
for the tiny build, and the MSVC CRT deprecation pragma carried over.

Builds and tests pass on all platforms (GNU full/tiny, MSVC).
